+++ src/main/java/com/devnity/devnity/domain/mapgakco/dto/mapgakco/response/MapgakcoResponse.java
@@ -27,6 +27,8 @@ public class MapgakcoResponse {
   @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
   private LocalDateTime deadline;
   @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
+  private LocalDateTime meetingAt;
+  @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
   private LocalDateTime createdAt;
   @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
   private LocalDateTime updatedAt;
