+++ src/main/java/com/devnity/devnity/domain/mapgakco/repository/mapgakcocomment/MapgakcoCommentRepository.java
@@ -1,4 +1,4 @@
-package com.devnity.devnity.domain.mapgakco.repository;
+package com.devnity.devnity.domain.mapgakco.repository.mapgakcocomment;
 
 import com.devnity.devnity.domain.mapgakco.entity.Mapgakco;
 import com.devnity.devnity.domain.mapgakco.entity.MapgakcoComment;
